Add styling changes and refactor time functions
